adapt SEn matrix and keep tests passing for SO3
